Add audio processing operations for filtering, spectral analysis, statistics, and temporal manipulation
- Implemented HighPassFilter, LowPassFilter, and AWeighting classes for audio filtering in `filters.py`.
- Added FFT, IFFT, STFT, ISTFT, Welch, NOctSpectrum, NOctSynthesis, Coherence, CSD, and TransferFunction classes for spectral analysis in `spectral.py`.
- Introduced ABS, Power, Sum, Mean, and ChannelDifference classes for statistical operations in `stats.py`.
- Created ReSampling, Trim, and RmsTrend classes for temporal manipulation in `temporal.py`.
- Registered all new operations for use in the audio processing pipeline.
