+++ wandas/processing/base.py
@@ -1,0 +1,126 @@
+import inspect
+import logging
+from typing import Any, ClassVar, Generic, TypeVar
+
+import dask
+import dask.array as da
+from dask.array.core import Array as DaArray
+
+from wandas.utils.types import NDArrayComplex, NDArrayReal
+
+logger = logging.getLogger(__name__)
+
+_da_from_delayed = da.from_delayed  # type: ignore [unused-ignore]
+
+# Define TypeVars for input and output array types
+InputArrayType = TypeVar("InputArrayType", NDArrayReal, NDArrayComplex)
+OutputArrayType = TypeVar("OutputArrayType", NDArrayReal, NDArrayComplex)
+
+
+class AudioOperation(Generic[InputArrayType, OutputArrayType]):
+    """Abstract base class for audio processing operations."""
+
+    # Class variable: operation name
+    name: ClassVar[str]
+
+    def __init__(self, sampling_rate: float, **params: Any):
+        """
+        Initialize AudioOperation.
+
+        Parameters
+        ----------
+        sampling_rate : float
+            Sampling rate (Hz)
+        **params : Any
+            Operation-specific parameters
+        """
+        self.sampling_rate = sampling_rate
+        self.params = params
+
+        # Validate parameters during initialization
+        self.validate_params()
+
+        # Create processor function (lazy initialization possible)
+        self._setup_processor()
+
+        logger.debug(
+            f"Initialized {self.__class__.__name__} operation with params: {params}"
+        )
+
+    def validate_params(self) -> None:
+        """Validate parameters (raises exception if invalid)"""
+        pass
+
+    def _setup_processor(self) -> None:
+        """Set up processor function (implemented by subclasses)"""
+        pass
+
+    def _process_array(self, x: InputArrayType) -> OutputArrayType:
+        """Processing function (implemented by subclasses)"""
+        # Default is no-op function
+        raise NotImplementedError("Subclasses must implement this method.")
+
+    @dask.delayed  # type: ignore [misc, unused-ignore]
+    def process_array(self, x: InputArrayType) -> OutputArrayType:
+        """Processing function wrapped with @dask.delayed"""
+        # Default is no-op function
+        logger.debug(f"Default process operation on data with shape: {x.shape}")
+        return self._process_array(x)
+
+    def calculate_output_shape(self, input_shape: tuple[int, ...]) -> tuple[int, ...]:
+        """
+        Calculate output data shape after operation (implemented by subclasses)
+
+        Parameters
+        ----------
+        input_shape : tuple
+            Input data shape
+
+        Returns
+        -------
+        tuple
+            Output data shape
+        """
+        raise NotImplementedError("Subclasses must implement this method.")
+
+    def process(self, data: DaArray) -> DaArray:
+        """
+        Execute operation and return result
+        data shape is (channels, samples)
+        """
+        # Add task as delayed processing
+        logger.debug("Adding delayed operation to computation graph")
+        delayed_result = self.process_array(data)
+        # Convert delayed result to dask array and return
+        output_shape = self.calculate_output_shape(data.shape)
+        return _da_from_delayed(delayed_result, shape=output_shape, dtype=data.dtype)
+
+
+# Automatically collect operation types and corresponding classes
+_OPERATION_REGISTRY: dict[str, type[AudioOperation[Any, Any]]] = {}
+
+
+def register_operation(operation_class: type) -> None:
+    """Register a new operation type"""
+
+    if not issubclass(operation_class, AudioOperation):
+        raise TypeError("Strategy class must inherit from AudioOperation.")
+    if inspect.isabstract(operation_class):
+        raise TypeError("Cannot register abstract AudioOperation class.")
+
+    _OPERATION_REGISTRY[operation_class.name] = operation_class
+
+
+def get_operation(name: str) -> type[AudioOperation[Any, Any]]:
+    """Get operation class by name"""
+    if name not in _OPERATION_REGISTRY:
+        raise ValueError(f"Unknown operation type: {name}")
+    return _OPERATION_REGISTRY[name]
+
+
+def create_operation(
+    name: str, sampling_rate: float, **params: Any
+) -> AudioOperation[Any, Any]:
+    """Create operation instance from name and parameters"""
+    operation_class = get_operation(name)
+    return operation_class(sampling_rate, **params)
